+++ src/models/immich/ImmichTypes.ts
@@ -13,6 +13,13 @@ export interface ImmichAlbum {
     [key: string]: any;
 }
 
+export interface ImmichTag {
+    id: string;
+    value: string;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    [key: string]: any;
+}
+
 export interface ImmichShareResponse {
     assets?: ImmichAsset[];
     asset?: ImmichAsset;